+++ src/elements/conditions/orders/OrderSiteConditionRule.php
@@ -1,0 +1,63 @@
+<?php
+
+namespace craft\commerce\elements\conditions\orders;
+
+use Craft;
+use craft\base\conditions\BaseMultiSelectConditionRule;
+use craft\base\ElementInterface;
+use craft\commerce\elements\db\OrderQuery;
+use craft\commerce\elements\Order;
+use craft\elements\conditions\ElementConditionRuleInterface;
+use craft\helpers\ArrayHelper;
+use yii\db\QueryInterface;
+
+/**
+ * Order site condition rule.
+ *
+ * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
+ * @since 4.2.7
+ */
+class OrderSiteConditionRule extends BaseMultiSelectConditionRule implements ElementConditionRuleInterface
+{
+    /**
+     * @inheritdoc
+     */
+    public function getLabel(): string
+    {
+        return Craft::t('commerce', 'Order Site');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getExclusiveQueryParams(): array
+    {
+        return ['orderSiteId'];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function options(): array
+    {
+        return ArrayHelper::map(Craft::$app->getSites()->getAllSites(), 'id', 'name');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function modifyQuery(QueryInterface $query): void
+    {
+        /** @var OrderQuery $query */
+        $query->orderSiteId($this->paramValue());
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function matchElement(ElementInterface $element): bool
+    {
+        /** @var Order $element */
+        return $this->matchValue($element->orderSiteId);
+    }
+}